move fitting machinery to Fit/ subdirectory since there will be a lot more of it momentarily
