+++ CAFAna/CMakeLists.txt
@@ -81,6 +81,7 @@ add_subdirectory(Cuts)
 add_subdirectory(Decomp)
 add_subdirectory(Experiment)
 add_subdirectory(Extrap)
+add_subdirectory(Fit)
 add_subdirectory(Prediction)
 add_subdirectory(Systs)
 add_subdirectory(Vars)
