+++ CAFAna/Fit/Fit.h
@@ -2,9 +2,9 @@
 
 #include "CAFAna/Core/Spectrum.h"
 #include "CAFAna/Core/SystShifts.h"
+#include "CAFAna/Fit/MinuitFitter.h"
+#include "CAFAna/Fit/SeedList.h"
 #include "CAFAna/Prediction/IPrediction.h"
-
-#include "CAFAna/Analysis/SeedList.h"
 
 #include "Math/Minimizer.h"
